fix(dts-plugin): dev plugin should apply after fetchPromise resolved
