+++ .changeset/tame-numbers-sneeze.md
@@ -1,0 +1,5 @@
+---
+'@module-federation/dts-plugin': patch
+---
+
+fix(dts-plugin): dev plugin should apply after fetchPromise resolved
