+++ packages/dts-plugin/src/plugins/DtsPlugin.ts
@@ -1,7 +1,11 @@
-import type { Compiler, WebpackPluginInstance } from 'webpack';
-import { type moduleFederationPlugin } from '@module-federation/sdk';
 import { DevPlugin } from './DevPlugin';
-import { TypesPlugin } from './TypesPlugin';
+import { normalizeOptions } from '@module-federation/sdk';
+import { ConsumeTypesPlugin } from './ConsumeTypesPlugin';
+import { GenerateTypesPlugin } from './GenerateTypesPlugin';
+import { isTSProject } from '../core';
+
+import { type moduleFederationPlugin } from '@module-federation/sdk';
+import type { Compiler, WebpackPluginInstance } from 'webpack';
 
 export class DtsPlugin implements WebpackPluginInstance {
   options: moduleFederationPlugin.ModuleFederationPluginOptions;
@@ -11,7 +15,49 @@ export class DtsPlugin implements WebpackPluginInstance {
 
   apply(compiler: Compiler) {
     const { options } = this;
-    new DevPlugin(options).apply(compiler);
-    new TypesPlugin(options).apply(compiler);
+
+    const defaultGenerateTypes = {
+      generateAPITypes: true,
+      compileInChildProcess: true,
+      abortOnError: false,
+      extractThirdParty: true,
+      extractRemoteTypes: true,
+    };
+    const defaultConsumeTypes = { abortOnError: false, consumeAPITypes: true };
+    const normalizedDtsOptions =
+      normalizeOptions<moduleFederationPlugin.PluginDtsOptions>(
+        isTSProject(options.dts, compiler.context),
+        {
+          generateTypes: defaultGenerateTypes,
+          consumeTypes: defaultConsumeTypes,
+          extraOptions: {},
+        },
+        'mfOptions.dts',
+      )(options.dts);
+
+    if (typeof normalizedDtsOptions !== 'object') {
+      return;
+    }
+
+    let resolve;
+
+    const fetchTypesPromise: Promise<void> = new Promise((res, rej) => {
+      resolve = res;
+    });
+
+    new DevPlugin(options, fetchTypesPromise).apply(compiler);
+
+    new GenerateTypesPlugin(
+      options,
+      normalizedDtsOptions,
+      defaultGenerateTypes,
+      fetchTypesPromise,
+    ).apply(compiler);
+    new ConsumeTypesPlugin(
+      options,
+      normalizedDtsOptions,
+      defaultConsumeTypes,
+      resolve,
+    ).apply(compiler);
   }
 }
